refactor(analysis): extract private helpers to _helpers.py
Move 7 private helper functions from deep_analysis_pipeline.py into a new
_helpers.py module. Re-export them from the pipeline for backwards compat.
Adds smoke tests for the extracted helpers.
